can `npm test` debug socket
